+++ R/RcppExports.R
@@ -17,11 +17,11 @@ loglik <- function(data, SI, para, n_inf, n_sus, with_rm, sep1, sep2) {
     .Call(`_hhdynamics_loglik`, data, SI, para, n_inf, n_sus, with_rm, sep1, sep2)
 }
 
-all_update <- function(data, dataorg, SI, para, member, loglik1, loglik2, n_inf, n_sus, with_rm, sep1, sep2) {
-    .Call(`_hhdynamics_all_update`, data, dataorg, SI, para, member, loglik1, loglik2, n_inf, n_sus, with_rm, sep1, sep2)
+all_update <- function(data, dataorg, SI, para, member, loglik1, loglik2, n_inf, n_sus, with_rm, sep1, sep2, factor_group, n_levels_vec) {
+    .Call(`_hhdynamics_all_update`, data, dataorg, SI, para, member, loglik1, loglik2, n_inf, n_sus, with_rm, sep1, sep2, factor_group, n_levels_vec)
 }
 
-mcmc <- function(data1, SI, mcmc_n, burnin, thinning, int_para, move, sigma, n_inf, n_sus, with_rm, sep1, sep2) {
-    .Call(`_hhdynamics_mcmc`, data1, SI, mcmc_n, burnin, thinning, int_para, move, sigma, n_inf, n_sus, with_rm, sep1, sep2)
+mcmc <- function(data1, SI, mcmc_n, burnin, thinning, int_para, move, sigma, n_inf, n_sus, with_rm, sep1, sep2, factor_group, n_levels_vec) {
+    .Call(`_hhdynamics_mcmc`, data1, SI, mcmc_n, burnin, thinning, int_para, move, sigma, n_inf, n_sus, with_rm, sep1, sep2, factor_group, n_levels_vec)
 }