+++ src/RcppExports.cpp
@@ -71,8 +71,8 @@ BEGIN_RCPP
 END_RCPP
 }
 // all_update
-List all_update(NumericMatrix data, NumericMatrix dataorg, NumericVector SI, NumericVector para, int member, NumericMatrix loglik1, NumericMatrix loglik2, int n_inf, int n_sus, int with_rm, int sep1, int sep2);
-RcppExport SEXP _hhdynamics_all_update(SEXP dataSEXP, SEXP dataorgSEXP, SEXP SISEXP, SEXP paraSEXP, SEXP memberSEXP, SEXP loglik1SEXP, SEXP loglik2SEXP, SEXP n_infSEXP, SEXP n_susSEXP, SEXP with_rmSEXP, SEXP sep1SEXP, SEXP sep2SEXP) {
+List all_update(NumericMatrix data, NumericMatrix dataorg, NumericVector SI, NumericVector para, int member, NumericMatrix loglik1, NumericMatrix loglik2, int n_inf, int n_sus, int with_rm, int sep1, int sep2, IntegerVector factor_group, IntegerVector n_levels_vec);
+RcppExport SEXP _hhdynamics_all_update(SEXP dataSEXP, SEXP dataorgSEXP, SEXP SISEXP, SEXP paraSEXP, SEXP memberSEXP, SEXP loglik1SEXP, SEXP loglik2SEXP, SEXP n_infSEXP, SEXP n_susSEXP, SEXP with_rmSEXP, SEXP sep1SEXP, SEXP sep2SEXP, SEXP factor_groupSEXP, SEXP n_levels_vecSEXP) {
 BEGIN_RCPP
     Rcpp::RObject rcpp_result_gen;
     Rcpp::RNGScope rcpp_rngScope_gen;
@@ -88,13 +88,15 @@ BEGIN_RCPP
     Rcpp::traits::input_parameter< int >::type with_rm(with_rmSEXP);
     Rcpp::traits::input_parameter< int >::type sep1(sep1SEXP);
     Rcpp::traits::input_parameter< int >::type sep2(sep2SEXP);
-    rcpp_result_gen = Rcpp::wrap(all_update(data, dataorg, SI, para, member, loglik1, loglik2, n_inf, n_sus, with_rm, sep1, sep2));
+    Rcpp::traits::input_parameter< IntegerVector >::type factor_group(factor_groupSEXP);
+    Rcpp::traits::input_parameter< IntegerVector >::type n_levels_vec(n_levels_vecSEXP);
+    rcpp_result_gen = Rcpp::wrap(all_update(data, dataorg, SI, para, member, loglik1, loglik2, n_inf, n_sus, with_rm, sep1, sep2, factor_group, n_levels_vec));
     return rcpp_result_gen;
 END_RCPP
 }
 // mcmc
-List mcmc(NumericMatrix data1, NumericVector SI, int mcmc_n, int burnin, int thinning, NumericVector int_para, NumericVector move, NumericVector sigma, int n_inf, int n_sus, int with_rm, int sep1, int sep2);
-RcppExport SEXP _hhdynamics_mcmc(SEXP data1SEXP, SEXP SISEXP, SEXP mcmc_nSEXP, SEXP burninSEXP, SEXP thinningSEXP, SEXP int_paraSEXP, SEXP moveSEXP, SEXP sigmaSEXP, SEXP n_infSEXP, SEXP n_susSEXP, SEXP with_rmSEXP, SEXP sep1SEXP, SEXP sep2SEXP) {
+List mcmc(NumericMatrix data1, NumericVector SI, int mcmc_n, int burnin, int thinning, NumericVector int_para, NumericVector move, NumericVector sigma, int n_inf, int n_sus, int with_rm, int sep1, int sep2, IntegerVector factor_group, IntegerVector n_levels_vec);
+RcppExport SEXP _hhdynamics_mcmc(SEXP data1SEXP, SEXP SISEXP, SEXP mcmc_nSEXP, SEXP burninSEXP, SEXP thinningSEXP, SEXP int_paraSEXP, SEXP moveSEXP, SEXP sigmaSEXP, SEXP n_infSEXP, SEXP n_susSEXP, SEXP with_rmSEXP, SEXP sep1SEXP, SEXP sep2SEXP, SEXP factor_groupSEXP, SEXP n_levels_vecSEXP) {
 BEGIN_RCPP
     Rcpp::RObject rcpp_result_gen;
     Rcpp::RNGScope rcpp_rngScope_gen;
@@ -111,7 +113,9 @@ BEGIN_RCPP
     Rcpp::traits::input_parameter< int >::type with_rm(with_rmSEXP);
     Rcpp::traits::input_parameter< int >::type sep1(sep1SEXP);
     Rcpp::traits::input_parameter< int >::type sep2(sep2SEXP);
-    rcpp_result_gen = Rcpp::wrap(mcmc(data1, SI, mcmc_n, burnin, thinning, int_para, move, sigma, n_inf, n_sus, with_rm, sep1, sep2));
+    Rcpp::traits::input_parameter< IntegerVector >::type factor_group(factor_groupSEXP);
+    Rcpp::traits::input_parameter< IntegerVector >::type n_levels_vec(n_levels_vecSEXP);
+    rcpp_result_gen = Rcpp::wrap(mcmc(data1, SI, mcmc_n, burnin, thinning, int_para, move, sigma, n_inf, n_sus, with_rm, sep1, sep2, factor_group, n_levels_vec));
     return rcpp_result_gen;
 END_RCPP
 }
@@ -121,8 +125,8 @@ static const R_CallMethodDef CallEntries[] = {
     {"_hhdynamics_prior_loglik", (DL_FUNC) &_hhdynamics_prior_loglik, 1},
     {"_hhdynamics_sim_data", (DL_FUNC) &_hhdynamics_sim_data, 8},
     {"_hhdynamics_loglik", (DL_FUNC) &_hhdynamics_loglik, 8},
-    {"_hhdynamics_all_update", (DL_FUNC) &_hhdynamics_all_update, 12},
-    {"_hhdynamics_mcmc", (DL_FUNC) &_hhdynamics_mcmc, 13},
+    {"_hhdynamics_all_update", (DL_FUNC) &_hhdynamics_all_update, 14},
+    {"_hhdynamics_mcmc", (DL_FUNC) &_hhdynamics_mcmc, 15},
     {NULL, NULL, 0}
 };